initial offspring index
