+++ app/controllers/rooms_controller.rb
@@ -1,0 +1,5 @@
+class RoomsController < ApplicationController
+  def index
+    @rooms = Room.all
+  end
+end
