+++ app/helpers/user/offsprings_helper.rb
@@ -1,0 +1,2 @@
+module User::OffspringsHelper
+end
