+++ app/views/layouts/_admin_header.html.haml
@@ -6,7 +6,7 @@
       %span.icon-bar
       %span.icon-bar
     %a.navbar-brand{href: "/"}
-      = image_tag("brand.svg", height: "64", alt: (t 'application.name').to_s)
+      = image_tag("brand.svg", height: "128", alt: (t 'application.name').to_s)
   .collapse.navbar-collapse.navbar-collapse-1
     %ul.nav.navbar-nav.navbar-utility
       %li
