+++ app/views/rooms/_shifts.html.haml
@@ -1,0 +1,30 @@
+/ Rendering shifts in order
+- if Rails.application.config.include_weekends
+  .col-md-7
+    .row
+      - (0..3).each do |day|
+        .col-md-3
+          %ul.list-group
+            %li.list-group-item.list-group-item-heading.active= week_day_convert(day)
+            - room.shifts.where(day_of_week: day).order("start_time").each do |shift|
+              %li.list-group-item-text
+                #{shift.start_time} - #{shift.end_time}, #{pluralize shift.sites_available, t('application.models.shift.site')}
+  .col-md-5
+    .row
+      - (4..6).each do |day|
+        .col-md-4
+          %ul.list-group
+            %li.list-group-item.list-group-item-heading.active= week_day_convert(day)
+            - room.shifts.where(day_of_week: day).order("start_time").each do |shift|
+              %li.list-group-item-text
+                #{shift.start_time} - #{shift.end_time}, #{pluralize shift.sites_available, t("application.models.shift.site")}
+- else
+  .col-md-12
+    .row
+      - (0..4).each do |day|
+        .col-md-2
+          %ul.list-group
+            %li.list-group-item.list-group-item-heading.active= week_day_convert(day)
+            - room.shifts.where(day_of_week: day).order("start_time").each do |shift|
+              %li.list-group-item-text
+                #{shift.start_time} - #{shift.end_time}, #{pluralize shift.sites_available, t('application.models.shift.site')}
