+++ app/views/rooms/index.html.haml
@@ -1,0 +1,38 @@
+- provide(:title, (t '.room_index').to_s)
+%h1= (t '.room_index')
+.row
+  %h2= (t "application.models.room.there_are_no_rooms_yet") if @rooms.empty?
+  = render @rooms
+
+- content_for :javascripts do
+  :javascript
+    (function($) {
+      $(document).ready(function() {
+        // Upon clicking the find button, show the find dropdown content
+        $(".btn-find").click(function () {
+          $(".find-pf-dropdown-container").toggle();
+        });
+        // Upon clicking the find close button, hide the find dropdown content
+        $(".btn-find-close").click(function () {
+          $(".find-pf-dropdown-container").hide();
+        });
+      });
+    })(jQuery);
+
+     $(document).ready(function () {
+      // Row Checkbox Selection
+      $("input[type='checkbox']").change(function (e) {
+        if ($(this).is(":checked")) {
+          $(this).closest('.list-group-item').addClass("active");
+        } else {
+          $(this).closest('.list-group-item').removeClass("active");
+        }
+      });
+      // toggle dropdown menu
+      $('.list-view-pf-actions').on('show.bs.dropdown', function () {
+        var $this = $(this);
+        var $dropdown = $this.find('.dropdown');
+        var space = $(window).height() - $dropdown[0].getBoundingClientRect().top - $this.find('.dropdown-menu').outerHeight(true);
+        $dropdown.toggleClass('dropup', space < 10);
+      });
+    });
