+++ app/views/user/offsprings/_add_offspring.html.haml
@@ -1,0 +1,3 @@
+%row
+  %p
+    = link_to (t ".add_offspring"), new_user_offspring_path(@user), class: "btn btn-lg btn-primary"
