+++ app/views/user/offsprings/_offspring.html.haml
@@ -1,0 +1,29 @@
+.list-group.list-view-pf-equalized-column
+  .list-group-item.list-view-pf-stacked
+    .list-view-pf-actions
+      %button.btn.btn-default= link_to (t '.assign'), "#"
+      .dropdown.pull-right.dropdown-kebab-pf
+        %button#dropdownKebabRight3.btn.btn-link.dropdown-toggle{"aria-expanded": "true", "aria-haspopup": "true", "data-toggle": "dropdown", type: "button"}
+          %span.fa.fa-ellipsis-v
+        %ul.dropdown-menu.dropdown-menu-right{"aria-labelledby" => "dropdownKebabRight3"}
+          %li
+            %a{:href => "#"} Edit
+          %li.divider{:role => "separator"}
+          %li
+            %a{:href => "#"} Delete
+
+    .list-view-pf-main-info
+      .list-view-pf-body
+        .list-view-pf-description
+          .list-group-item-heading
+            = offspring.first_name
+          .list-group-item-text
+            = offspring.last_name
+        .list-view-pf-additional-info
+          .list-view-pf-additional-info-item.list-view-pf-additional-info-item
+            %strong= offspring.assignment
+            Aula donde va todo esto
+          .list-view-pf-additional-info-item.list-view-pf-additional-info-item
+            %strong 35%
+            Turno asignado
+
