+++ app/views/user/offsprings/create.html.haml
@@ -1,0 +1,2 @@
+%h1 User::Offsprings#create
+%p Find me in app/views/user/offsprings/create.html.haml
