+++ app/views/user/offsprings/destroy.html.haml
@@ -1,0 +1,2 @@
+%h1 User::Offsprings#destroy
+%p Find me in app/views/user/offsprings/destroy.html.haml
