+++ app/views/user/offsprings/edit.html.haml
@@ -1,0 +1,2 @@
+%h1 User::Offsprings#edit
+%p Find me in app/views/user/offsprings/edit.html.haml
