+++ app/views/user/offsprings/new.html.haml
@@ -1,0 +1,2 @@
+%h1 User::Offsprings#new
+%p Find me in app/views/user/offsprings/new.html.haml
