+++ app/views/user/offsprings/show.html.haml
@@ -1,0 +1,2 @@
+%h1 User::Offsprings#show
+%p Find me in app/views/user/offsprings/show.html.haml
