+++ app/views/user/offsprings/update.html.haml
@@ -1,0 +1,2 @@
+%h1 User::Offsprings#update
+%p Find me in app/views/user/offsprings/update.html.haml
