+++ spec/controllers/rooms_controller_spec.rb
@@ -1,0 +1,4 @@
+require 'rails_helper'
+
+RSpec.describe RoomsController, type: :controller do
+end
